Removed 5V linear regulator because it generated too much heat.
The 5V linear regulator generated a ton of heat right above the Pi
processor, which caused it to thermally-overload and reset. So, we won't
try to power the Pi from this board for now. In the future, we could use
a more efficient switching supply that wouldn't generate as much heat.
But, we can separately power the Pi via USB for now.